refactored benchmark
